move auth module to modules folder
