+++ Src/Campus.Master.API/UI/src/app/app-routing.module.ts
@@ -1,11 +1,11 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { UserPageLayoutComponent } from './views/user-page-layout/user-page-layout.component';
-import { CanActivateDashboardGuard } from './core/guards/can-activate-dashboard.guard';
+import { CanActivateDashboardGuard } from '@core/guards/can-activate-dashboard.guard';
 
 
 const routes: Routes = [
-  { path: '', loadChildren: () => import('./auth/auth.module').then(m => m.AuthModule) },
+  { path: '', loadChildren: () => import('./modules/auth/auth.module').then(m => m.AuthModule) },
   {
     path: 'campus', component: UserPageLayoutComponent, canActivate: [CanActivateDashboardGuard]
   }
