+++ Src/Campus.Master.API/UI/src/app/modules/auth/auth-routing.module.ts
