+++ Src/Campus.Master.API/UI/src/app/modules/auth/auth.module.ts
@@ -3,8 +3,8 @@ import { NgModule } from '@angular/core';
 import { AuthRoutingModule } from './auth-routing.module';
 import { RegistrationComponent } from './registration/registration.component';
 import { SingInComponent } from './sing-in/sing-in.component';
-import { SharedModule } from '../shared/shared.module';
 import { RegistrationService } from './shared/services/registration.service';
+import { SharedModule } from '@shared/shared.module';
 
 @NgModule({
   declarations: [SingInComponent, RegistrationComponent],
