+++ Src/Campus.Master.API/UI/src/app/modules/auth/registration/registration.component.sass
