+++ Src/Campus.Master.API/UI/src/app/modules/auth/shared/models/authenticated-user.ts
