+++ Src/Campus.Master.API/UI/src/app/modules/auth/sing-in/sing-in.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
-import { SignInService } from '../../core/sign-in.service';
+import { SignInService } from '@core/sign-in.service';
 import { Router } from '@angular/router';
 import { ToastrService } from 'ngx-toastr';
 import { Subscription, zip } from 'rxjs';
